Refactor CcapExpeditedEligibilityParserTest to use smaller test slice
